==> R/Part_3_Table_22_23_24.R <==
library(dplyr)

# install CSCI package if needed
# devtools::install_github('SCCWRP/CSCI')

## Traditional reference
csci_df <- CSCI::refsamples
asci_df <- readr::read_csv("data-raw/Part_1_asci_caldata.csv") |>
  rename_with(.fn = tolower) |>
  select(-...1) 


## Intermittent reference
nonperen_df <- readr::read_csv("data-raw/Part_1_Andy_Nonperen_data_ASCI.csv") |>
  mutate(
    masterid = StationCode,
    sampledate = lubridate::mdy(SampleDate),
    Region = if_else(RB %in% c(4, 7, 8, 9), "Southern", "Northern")
  )

sample_exclusion <- readr::read_csv('data-raw/Part_1_sample_metadata_review_07152024.csv') |>
  filter(!Include) |>
  mutate(sampledate = lubridate::mdy(SampleDate))

nonperen_df <- nonperen_df |>
  anti_join(sample_exclusion, by = c("StationCode", "sampledate"))


## Valley floor
test_data_gis <- readr::read_csv("data-raw/Part_1_test_data_gis.csv") |>
  mutate(logwsa = log10(area_sqkm)) |>
  group_by(masterid) |>
  slice_head(n = 1) |>
  ungroup() |>
  rename_with(.fn = tolower)

test_data_scores <- readr::read_csv("data-raw/Part_1_test_data_scores.csv") |>
  rename_with(.fn = tolower) |>
  inner_join(test_data_gis |> select(masterid, studyarea))

## Constructed streams and modified classes
chan_df <- readr::read_csv("data-raw/Part_2_combined_df_withdata.csv") |>
  select(masterid, class_do, Watershed, CSCI, ASCI_H, ASCI_D)

mydf <-
  ##Ref sites
  #CSCI
  csci_df |>
  transmute(
    masterid = StationCode, 
    CSCI,
    Type = "Reference"
  ) |>
  #ASCI
  full_join(
    asci_df |>
      filter(sitesetsample2 == 'Reference') |>
      transmute(masterid = stationcode, ASCI_D = d_asci, ASCI_H = h_asci, Type = "Reference")
  ) |>
  #Intermittent
  bind_rows(
    nonperen_df |> 
      filter(Flow_SOP == "RFI", Region == "Northern") |>
      transmute(masterid = StationCode, CSCI, ASCI_D, ASCI_H, Type = "Intermittent_NorCal"),
    nonperen_df |> 
      filter(Flow_SOP == "RFI", Region == "Southern") |>
      transmute(masterid = StationCode, CSCI, ASCI_D, ASCI_H, Type = "Intermittent_SoCal"),
    #General
    test_data_scores |> 
      inner_join(
        test_data_gis |> 
          select(masterid, studyarea) |>
          filter(studyarea == "Central Valley") |>
          group_by(masterid) |>
          slice_head(n=1) |>
          ungroup()
      ) |>
      transmute(masterid, CSCI = csci, ASCI_D = asci_d, ASCI_H = asci_h, Type = "CVFloor"),
    #Ambiguous 
    chan_df |>
      filter(Watershed == "Ambiguous") |>
      select(-class_do, -Watershed) |>
      mutate(Type = "Ambiguous"),
    #Modified
    chan_df |>
      filter(!is.na(class_do)) |>
      filter(class_do != "Natural") |>
      transmute(masterid, CSCI, ASCI_H, ASCI_D, Type = class_do),
    #Modified AND CVF 
    chan_df |>
      filter(
        !is.na(class_do),
        class_do!="Natural",
        masterid %in% test_data_gis$masterid[test_data_gis$studyarea == "Central Valley"]
      ) |>
      transmute(masterid, CSCI, ASCI_H, ASCI_D, Type = paste0(class_do, "_CVF")),
    chan_df |>
      filter(Watershed == "Ambiguous") |>
      mutate(RB = stringr::str_sub(masterid, 1, 1)) |>
      filter(RB == 5) |>
      transmute(masterid, CSCI, ASCI_H, ASCI_D, Type = ("CC_CVF"))
  ) |>
  #Reformat
  tidyr::pivot_longer(
    cols = c("CSCI","ASCI_D","ASCI_H"), values_drop_na = T, 
    names_to = "Index", values_to = "Score"
  ) |>
  mutate(
    Type2 = if_else(
      Type %in% c("Reference", "Intermittent_SoCal", "Intermittent_NorCal"),
      "Reference",
      "Best observed"
    ),
    Type = factor(
      Type,
      levels = c(
        "Ambiguous", "CC_CVF", "CVFloor", "Hard bottom", "Hard bottom_CVF", 
        "Intermittent_NorCal", "Intermittent_SoCal", "Reference", 
        "Soft bottom-0 hard sides", "Soft bottom-0 hard sides_CVF", 
        "Soft bottom-1 hard side", "Soft bottom-2 hard sides"
      ),
      labels = c(
        "CC", "CC_CVF","CVF", "HB", "HB_CVF",
        "RFI-N","RFI-S", "Reference",
        "SB0","SB0_CVF", "SB1","SB2"
      )
    ),
    Type = factor(
      Type, 
      levels = c(
        "Reference", "RFI-N", "RFI-S", "CVF", "SB0", "SB0_CVF", 
        "SB1", "SB2", "HB", "HB_CVF", "CC", "CC_CVF"
      )
    )
  ) |>
  arrange(Type2, Type, Index, Score) 


## Table 22 ####

bi_thresholds_table <- mydf |>
  group_by(masterid, Type2, Type, Index) |>
  summarize(value = mean(Score, na.rm = T)) |>
  ungroup() |>
  tidyr::pivot_wider(names_from = "Index", values_from = "value") |>
  tidyr::pivot_longer(cols = ASCI_D:CSCI, names_to = "Index", values_to = "value") |>
  group_by(Type2, Type, Index) |>
  summarize(
    n_sites = sum(!is.na(value)),
    qe01 = quantile(value, probs = c(0.01), na.rm = T),
    qe10 = quantile(value, probs = c(0.1), na.rm = T),
    qe30 = quantile(value, probs = c(0.3), na.rm = T),
    qe70 = quantile(value, probs = c(0.7), na.rm = T),
    qe90 = quantile(value, probs = c(0.9), na.rm = T),
    qe99 = quantile(value, probs = c(0.99), na.rm = T),
    MeanScore = mean(value, na.rm = T),
    SDScore = sd(value, na.rm = T),
    qn01 = qnorm(p = 0.01, mean = MeanScore, sd = SDScore, lower.tail = T),
    qn10 = qnorm(p = 0.10, mean = MeanScore, sd = SDScore, lower.tail = T),
    qn30 = qnorm(p = 0.30, mean = MeanScore, sd = SDScore, lower.tail = T),
    qn70 = qnorm(p = 0.70, mean = MeanScore, sd = SDScore, lower.tail = T),
    qn90 = qnorm(p = 0.90, mean = MeanScore, sd = SDScore, lower.tail = T),
    qn99 = qnorm(p = 0.99, mean = MeanScore, sd = SDScore, lower.tail = T)
  ) |>
  ungroup() |>
  arrange(Type)  |>
  transmute(
    `Threshold type` = Type2,
    Population = if_else(Type == 'Reference', 'Wadeable (standard)', Type), 
    Index = Index, 
    StandardUsageSupported = case_when(
      Type == "Reference" ~ "Yes",
      Type == "RFI-N" & Index %in% c("ASCI_D", "ASCI_H") ~ "Yes",
      Type == "RFI-N" & Index %in% c("CSCI") ~ "No",
      Type == "RFI-S" ~ "Yes",
      Type == "CVF" & Index %in% c("ASCI_D", "ASCI_H") ~ "Yes",
      Type == "CVF" & Index %in% c("CSCI") ~ "No",
      Type == "SB0" & Index %in% c("ASCI_D", "ASCI_H") ~ "No",
      Type == "SB0" & Index %in% c("CSCI") ~ "Yes",
      Type == "SB1" ~ "Yes",
      Type == "SB2" & Index %in% c("ASCI_D", "ASCI_H") ~ "No",
      Type == "SB2" & Index %in% c("CSCI") ~ "Yes",
      Type == "HB" & Index %in% c("ASCI_D", "ASCI_H") ~ "Yes",
      Type == "HB" & Index %in% c("CSCI") ~ "No",
      (Type == "CC" | Type == "CC_CVF") & Index %in% c("ASCI_D", "ASCI_H") ~ "ND",
      (Type == "CC" | Type == "CC_CVF") & Index %in% c("CSCI") ~ "No",
      .default = "ND"
    ),
    n_sites = n_sites,
    High = case_when(
      Type2 == "Reference" & Type %in% c("RFI-N", "RFI-S") ~ qn30,
      Type2 == "Reference" & Type %in% c("Reference") & Index == "CSCI" ~ 0.92,
      Type2 == "Reference" & Type %in% c("Reference") & Index %in% c("ASCI_D", "ASCI_H") ~ 0.94,
      Type2 == "Best observed" ~ qe99,
      .default = -999
    ),
    Intermediate = case_when(
      Type2 == "Reference" & Type %in% c("RFI-N","RFI-S") ~ qn10,
      Type2 == "Reference" & Type %in% c("Reference") & Index == "CSCI" ~ 0.79,
      Type2 == "Reference" & Type %in% c("Reference") & Index %in% c("ASCI_D", "ASCI_H") ~ 0.86,
      Type2 == "Best observed" ~ qe90,
      .default = -999
    ),
    Low = case_when(
      Type2 == "Reference" & Type %in% c("RFI-N", "RFI-S") ~ qn01,
      Type2 == "Reference" & Type %in% c("Reference") & Index == "CSCI" ~ 0.63,
      Type2 == "Reference" & Type %in% c("Reference") & Index %in% c("ASCI_D", "ASCI_H") ~ 0.75,
      Type2 == "Best observed" ~ qe70,
      .default = -999
    )
  ) |>
  mutate(across(High:Low, .fns = function(x) round(x, 2)))

#Numbers from Theroux et al. 2022 or Mazor et al. 2016
bi_thresholds_table$n_sites[bi_thresholds_table$Population == 'Wadeable (standard)'] <- c(418, 418, 473) 

write.csv(bi_thresholds_table, "tables/Part_3_Table_22.csv", row.names = F)

## Table 23 ####
all_data_chem_phab <- readr::read_csv("data-raw/Part_3_all_data_chem_phab.csv")

# use eutrophication reference data from Mazor et al. 2022
bs_ref_data <- readr::read_csv("data-raw/Part_3_state_summary_r_table.csv") |>
  mutate(
    Type2 = "Reference",
    Type = factor("Reference", levels = levels(mydf$Type)),
    SiteStatus = NULL
  ) |>
  rename(
    qe01 = Q01, qe10 = Q10, qe30 = Q30,
    qe70 = Q70, qe90 = Q90, qe99 = Q99,
    n_sites = n, name = variable
  ) |>
  mutate(
    qe01 = if_else(name == "Ash_Free_Dry_Mass_mgPercm2", qe01 * 10, qe01), # convert to g/m2 just for AFDM
    qe10 = if_else(name == "Ash_Free_Dry_Mass_mgPercm2", qe10 * 10, qe10), 
    qe30 = if_else(name == "Ash_Free_Dry_Mass_mgPercm2", qe30 * 10, qe30), 
    qe70 = if_else(name == "Ash_Free_Dry_Mass_mgPercm2", qe70 * 10, qe70), 
    qe90 = if_else(name == "Ash_Free_Dry_Mass_mgPercm2", qe90 * 10, qe90), 
    qe99 = if_else(name == "Ash_Free_Dry_Mass_mgPercm2", qe99 * 10, qe99),
    name = case_when(
      name == "Ash_Free_Dry_Mass_mgPercm2" ~ "AFDM_g_m2",
      name == "Chlorophyll_a_mgPerm2" ~ "Chl-a_mg_m2",
      name == "Nitrogen_Total_mgPerL" ~ "TN",
      name == "Phosphorus_as_P_mgPerL" ~ "TP",
      .default = name
    )
  )

mydf_bs <- mydf |>
  select(masterid, Type, Type2) |>
  filter(Type != "Reference") |>
  unique() |>
  left_join(
    all_data_chem_phab |>
      select(masterid, sampledate, TN, TP, AFDM_g_m2, `Chl-a_mg_m2`, PCT_MAP)
  ) |>
  tidyr::pivot_longer(cols = c(TN, TP, AFDM_g_m2, `Chl-a_mg_m2`, PCT_MAP), values_drop_na = T) |>
  group_by(masterid, Type, Type2, name) |>
  summarize(value = mean(value)) |>
  ungroup() |>
  group_by(Type2, Type, name) |>
  summarize(
    n_sites = sum(!is.na(value)),
    qe01 = quantile(value, probs = c(0.01), na.rm = T),
    qe10 = quantile(value, probs = c(0.1), na.rm = T),
    qe30 = quantile(value, probs = c(0.3), na.rm = T),
    qe70 = quantile(value, probs = c(0.7), na.rm = T),
    qe90 = quantile(value, probs = c(0.9), na.rm = T),
    qe99 = quantile(value, probs = c(0.99), na.rm = T)
  ) |>
  ungroup() |>
  bind_rows(bs_ref_data) |>
  mutate(
    name = factor(name, levels = c("TN", "TP", "Chl-a_mg_m2", "AFDM_g_m2", "PCT_MAP"))
  ) |>
  arrange(name, desc(Type2), Type)

table_23 <- mydf_bs |>
  transmute(
    `Threshold type` = Type2,
    Population = if_else(Type == 'Reference', 'Wadeable (standard)', Type),
    Indicator = case_when(
      name == "AFDM_g_m2" ~ "AFDM (g/m2)",
      name == "Chl-a_mg_m2" ~ "Chl-a (mg/m2)",
      name == "PCT_MAP" ~ "% cover",
      name == "TN" ~ "TN (mg/L)",
      name == "TP" ~ "TP (mg/L)",
      .default = name
    ),
    Indicator = factor(Indicator, levels = c("TN (mg/L)", "TP (mg/L)", "Chl-a (mg/m2)", "AFDM (g/m2)", "% cover")),
    n = n_sites,
    High = case_when(
      Type2 == "Reference" ~ qe70,
      Type2 == "Best observed" ~ qe01,
      .default = -999
    ),
    Intermediate = case_when(
      Type2 == "Reference" ~ qe90,
      Type2 == "Best observed" ~ qe10,
      .default = -999
    ),
    Low = case_when(
      Type2 == "Reference" ~ qe99,
      Type2 == "Best observed" ~ qe30,
      .default = -999
    )
  ) |>
  filter(!(Population %in% c("SB0_CVF", "HB_CVF", "CC_CVF"))) |>
  mutate(across(High:Low, .fns = function(x) {
      case_when(
        Indicator == "AFDM (g/m2)" ~ round(x, 1),
        Indicator == "TN (mg/L)" ~ round(x, 3),
        Indicator == "TP (mg/L)" ~ round(x, 3),
        Indicator == "Chl-a (mg/m2)" ~ round(x, 1),
        Indicator == "% cover" ~ round(x)
      )
    })
  ) |>
  select(Indicator, `Threshold type`, `Stream class` = Population, n, High, Intermediate, Low) |>
  arrange(Indicator, desc(`Threshold type`)) 

write.csv(table_23, "tables/Part_3_Table_23.csv", row.names = F)

## Table 24 ####

mazor2022_thresholds <- readr::read_csv("data-raw/Part_3_model.summary_thresholds_woInorganics.csv") |>
  filter(BIgoal %in% c("Ref01", "Ref10", "Ref30"), Stratum == "California") |>
  transmute(
    Class = "Wadeable streams",
    Index = Response,
    BSPretty,
    Stringency = case_when(
      BIgoal == "Ref30" ~ "High",
      BIgoal == "Ref10" ~ "Intermediate",
      BIgoal == "Ref01" ~ "Low",
      .default = "x"
    ),
    Threshold = p80
  ) |>
  tidyr::pivot_wider(names_from = BSPretty, values_from = Threshold) |>
  mutate(
    AFDM = AFDM * 10 # convert from mg/cm2 to g/m2
  ) |>
  filter(Stringency == 'Intermediate') |>
  transmute(
    Population = Class, Index, StandardUsageSupported = "Yes", 
    value = c(0.86, 0.86, 0.79), AFDM, `Chl-a`, `Total N`, `% cover`, `Total P`
  )

mydf.c <- readr::read_csv("data-raw/Part_3_mydf.c.csv") |>
  mutate(
    PSA2 = if_else(PSA6c %in% c("NC", "SN"), "Wet", "Arid"),
    Ash_Free_Dry_Mass_gPerm2 = Ash_Free_Dry_Mass_mgPercm2 * 10, # convert from mg/cm2 to g/m2
  )

asci.df.c <- readr::read_csv("data-raw/Part_3_asci.df.c.csv") |>
  mutate(
    PSA2 = if_else(PSA6c %in% c("NC", "SN"), "Wet", "Arid"),
    Ash_Free_Dry_Mass_gPerm2 = Ash_Free_Dry_Mass_mgPercm2 * 10, # convert from mg/cm2 to g/m2
  )

#Create vectors of modeling variables
#Biostiulatory variables
chem.varz <- c("Nitrogen_Total_mgPerL", "Phosphorus_as_P_mgPerL")
om.varz <- c("Chlorophyll_a_mgPerm2", "Ash_Free_Dry_Mass_gPerm2", "PCT_MAP")
bs.varz <- c(chem.varz, om.varz)
bs.varz_pretty <- c("Total N", "Total P", "Chl-a", "AFDM", "% cover") #This is for graphs and plots
bs.varz.df <- data.frame(BiostimVar = bs.varz, BSPretty = factor(bs.varz_pretty, levels = bs.varz_pretty))
mod.id.varz <- c("MasterID", "MasterDate", "MasterDate_rep", "DevSet", "SelectedSample", "PSA2", "New_Lat", "New_Long")

mod.dat.csci <- mydf.c |>
  select(all_of(mod.id.varz), CSCI, all_of(chem.varz), all_of(om.varz)) |>
  rename(IndexScore = CSCI) |>
  mutate(Index = "CSCI")
mod.dat.asci_d<- asci.df.c |>
  select(all_of(mod.id.varz), ASCI_D, all_of(chem.varz), all_of(om.varz)) |>
  rename(IndexScore = ASCI_D) |>
  mutate(Index = "ASCI_D")
mod.dat.asci_h <- asci.df.c |>
  select(all_of(mod.id.varz), ASCI_H, all_of(chem.varz), all_of(om.varz)) |>
  rename(IndexScore = ASCI_H) |>
  mutate(Index = "ASCI_H")

mod.dat <- bind_rows(mod.dat.csci, mod.dat.asci_d, mod.dat.asci_h)

#CREATE MODELS
length.out.x <- 1000

max_TN <- 3
max_TP <- 1.5
max_chl <- 300
max_afdm <- 400
max_cov <- 100

my.newdfs <- data.frame(
    Nitrogen_Total_mgPerL = seq(from = 0, to = max_TN, length.out = length.out.x),
    Phosphorus_as_P_mgPerL = seq(from = 0, to = max_TP, length.out = length.out.x),
    Chlorophyll_a_mgPerm2 = seq(from = 0, to = max_chl, length.out = length.out.x),
    Ash_Free_Dry_Mass_gPerm2 = seq(from = 0, to = max_afdm, length.out = length.out.x),
    PCT_MAP = seq(from = 0, to = max_cov, length.out = length.out.x)
  ) |>
  tidyr::pivot_longer(cols = everything(), names_to = "BiostimVar", values_to = "Biostim") |>
  tidyr::nest(.by = "BiostimVar")

model_df <- mod.dat |>
  filter(DevSet == "Cal", SelectedSample == "Selected") |>
  tidyr::pivot_longer(cols = Nitrogen_Total_mgPerL:PCT_MAP, names_to = "BiostimVar", values_to = "Biostim") |>
  na.omit() |>
  group_by(Index, BiostimVar) |>
  mutate(
    model = list(scam::scam(IndexScore ~ s(Biostim, bs = "mpd")))
  )

my.models.predictions <- model_df |>
  inner_join(my.newdfs) |>
  select(model, data) |>
  distinct() |>
  mutate(
    Fit = list(predict(first(model), newdata = first(data), type = "response", se = T)$fit),
    SE = list(predict(first(model), newdata = first(data), type = "response", se = T)$se.fit)
  ) |>
  tidyr::unnest(cols = c(data, Fit, SE))

table_24 <- bi_thresholds_table |>
  filter(!(Population %in% c("SB0_CVF", "HB_CVF", "CC_CVF", "RFI-N", "RFI-S"))) |>
  select(Index, Population, StandardUsageSupported, High, Intermediate, Low) |>
  tidyr::pivot_longer(cols = c(High, Intermediate, Low)) |>
  unique() |>
  mutate(value = round(value, 2)) |>
  tidyr::crossing(BiostimVar = bs.varz) |> 
  left_join(my.models.predictions, by = join_by(Index == Index, BiostimVar == BiostimVar, closest(value <= Fit))) |>
  filter(name == "Intermediate") |>
  select(-c(Fit, SE)) |>
  inner_join(bs.varz.df) |>
  tidyr::pivot_wider(
    names_from = BSPretty, values_from = Biostim, 
    id_cols = c(Population, Index, StandardUsageSupported, value)
  ) |>
  bind_rows(mazor2022_thresholds) |>
  mutate(
    Population = factor(
      Population,
      levels = c("Wadeable streams", "Wadeable (standard)", "CVF", "SB0", "SB1", "SB2", "HB", "CC"),
      labels = c(
        "Wadeable streams (logistic regression, Mazor et al. 2022)", 
        "Wadeable streams (SCAM, present study)", "CVF", "SB0", "SB1", "SB2", "HB", "CC"
      )
    ),
    `Total N` = if_else(`Total N` == max_TN, NA_real_, `Total N`) |> round(3),
    `Total P` = if_else(`Total P` == max_TP, NA_real_, `Total P`) |> round(3),
    `Chl-a` = if_else(`Chl-a` == max_chl, NA_real_, `Chl-a`) |> round(1),
    AFDM = if_else(AFDM == max_afdm, NA_real_, AFDM) |> round(1),
    `% cover` = if_else(`% cover` == max_cov, NA_real_, `% cover`) |> round()
  ) |>
  arrange(Population, Index) |>
  filter(!is.na(value)) |>
  select(
    `Stream class` = Population, Index, StandardUsageSupported, `Biointegrity goal` = value, 
    `Total N`, `Total P`, `Chl-a`, AFDM, `% cover`
  )

write.csv(table_24, "tables/Part_3_Table_24.csv", row.names = F)
